# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para rastreamento de geolocalização de usuários em eventos\nEsta API permite criar usuários, salvar posições geográficas e consultar usuários próximos",
        "title": "Geolocation Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/positions": {
            "post": {
                "description": "Salva uma nova posição geográfica para um usuário específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "positions"
                ],
                "summary": "Salvar posição do usuário",
                "parameters": [
                    {
                        "description": "Dados da posição",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SavePositionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Posição salva com sucesso",
                        "schema": {
                            "$ref": "#/definitions/usecase.SaveUserPositionResponse"
                        }
                    },
                    "400": {
                        "description": "Dados de posição inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/positions/nearby": {
            "get": {
                "description": "Busca usuários próximos a uma coordenada específica dentro de um raio determinado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "positions"
                ],
                "summary": "Buscar usuários próximos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuário que está buscando",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Latitude da posição de referência (-90 a 90)",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude da posição de referência (-180 a 180)",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Raio de busca em metros (1 a 50000)",
                        "name": "radius_meters",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Número máximo de resultados (padrão: 50)",
                        "name": "max_results",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários próximos",
                        "schema": {
                            "$ref": "#/definitions/usecase.FindNearbyUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Parâmetros de busca inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/positions/sector": {
            "get": {
                "description": "Busca todos os usuários que estão no mesmo setor geográfico de uma coordenada específica",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "positions"
                ],
                "summary": "Buscar usuários no mesmo setor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuário que está buscando",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Latitude da posição de referência (-90 a 90)",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude da posição de referência (-180 a 180)",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários no setor",
                        "schema": {
                            "$ref": "#/definitions/usecase.GetUsersInSectorResponse"
                        }
                    },
                    "400": {
                        "description": "Parâmetros de busca inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.SavePositionRequest": {
            "type": "object",
            "required": [
                "latitude",
                "longitude",
                "user_id"
            ],
            "properties": {
                "latitude": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90
                },
                "longitude": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "usecase.FindNearbyUsersResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "nearby_users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.NearbyUserResponse"
                    }
                },
                "search_center": {
                    "$ref": "#/definitions/usecase.NearbyUserResponse"
                },
                "total_found": {
                    "type": "integer"
                }
            }
        },
...
        "usecase.GetUsersInSectorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "requested_by": {
                    "$ref": "#/definitions/usecase.SectorUserResponse"
                },
                "sector_bounds": {
                    "$ref": "#/definitions/usecase.SectorBounds"
                },
                "sector_id": {
                    "type": "string"
                },
                "total_found": {
                    "type": "integer"
                },
                "users_in_sector": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.SectorUserResponse"
                    }
                }
            }
        },
        "usecase.NearbyUserResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "Ex: \"5m30s\"",
                    "type": "string"
                },
                "distance_meters": {
                    "type": "number"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "position_id": {
                    "type": "string"
                },
                "sector_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "usecase.SaveUserPositionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "position_id": {
                    "type": "string"
                },
                "sector_id": {
                    "type": "string"
                }
            }
        },
        "usecase.SectorBounds": {
            "type": "object",
            "properties": {
                "max_latitude": {
                    "type": "number"
                },
                "max_longitude": {
                    "type": "number"
                },
                "min_latitude": {
                    "type": "number"
                },
                "min_longitude": {
                    "type": "number"
                }
            }
        },
        "usecase.SectorUserResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "Ex: \"5m30s\"",
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "position_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        }
    },
...
}